refactor(codemode): split eval cell orchestration and prompt template into own modules
eval-tool.ts sat at 258 pure LOC; runEvalCell/executeCell move to src/tool/run-eval-cell.ts and EVAL_PROMPT_TEMPLATE to src/prompt/eval-prompt-template.ts ahead of the run-budget change. No behavior change.
